+++ CHANGELOG.md
@@ -58,6 +58,14 @@
 * Add EXPORT option to fees/fines Details. Refs UIU-1956.
 * Change Overdue loans and Claims reports icons in action menu. Refs UIU-2030.
 * Add plus-sign to create buttons in action menu and switch button order. Refs UIU-2031.
+* Calculate new expiration date not only from today. Refs UIU-2046.
+* Unable to select today's date for refund report. Refs UIU-2033.
+* Increment `@folio/stripes-cli` to `v2`. Refs UIU-2047.
+* Inconsistent behavior of payment action when transaction info is blank. Refs UIU-2048.
+* Fix wrong value on reset data for Refunds to process manually report (CSV). Refs UIU-2050.
+* Add Callout for Refund report. Refs UIU-2035.
+* Allow 0 as valid entry for Patron Block Limits. Refs UIU-1998.
+* Add "Save and close" button should be active immediately on refund report criteria modal. Refs UIU-2034.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)